Move test-class global vars into tests and expand.
It is absolutely crucial that tests do not run an
assumption about the order in which they are run.

Avoid doing this in the future.

Also, Test scope was limited. Expanding the scope
revealed that a hunter will move if their path
is blocked in by a vertically aligned obsaticle
and they have a valid move "up", but not so if
horizontally aligned set of obstacles, and it
has a valid move "left".